apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-multipart==0.0.20
openai==1.107.1
numpy==2.3.2
orjson==3.11.2
vaderSentiment==3.3.2
